replace 'react' with 'youdash' and add appAccess
Replaced reactHTMLFile with youdashHTML file and all references to react with youdash. We are preparing for more than react app.

Added hooks for /app in path, and $appAccess variable, but did not complete appAccess. The static folder is common in react apps, and we must prefix it with something so that url rewrite can correctly identify and replace references to static/... in files.